% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employees.R
\docType{data}
\name{employees}
\alias{employees}
\title{An employees dataset for demonstrating the functions in allin1}
\format{
\subsection{\code{employees}}{

A data frame with 5 rows and 3 columns:
\describe{
\item{name}{A string denoting the name of employee}
\item{hours_worked}{Integer for hours worked}
\item{hourly_wage}{Integer for hourly_wage}
}
}
}
\usage{
employees
}
\description{
An employees dataset for demonstrating the functions in allin1
}
\keyword{datasets}
