% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_percent.R
\name{top_n_percent}
\alias{top_n_percent}
\title{Extract the top percentage of rows based on a column}
\usage{
top_n_percent(.data, col, percent)
}
\arguments{
\item{.data}{a dataframe}

\item{col}{Column name to sort by}

\item{percent}{number between 0 and 100 to specify the percentage}
}
\value{
A filtered and sorted dataframe
}
\description{
Extract the top percentage of rows based on a column
}
