% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_rows.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop completely empty rows}
\usage{
drop_na_rows(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
A dataframe with empty rows removed
}
\description{
Drop completely empty rows
}
\examples{
clean_employees <-employees |> drop_na_rows
clean_employees
}
