% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mutate.R
\name{sort_mutate}
\alias{sort_mutate}
\title{Applies sort_mutate}
\usage{
sort_mutate(.data, expr, desc = FALSE)
}
\arguments{
\item{.data}{dataframe}

\item{expr}{expression, mathematical formula for your mutate}

\item{desc}{automatically set to FALSE}
}
\value{
a table with mutated new column
}
\description{
Applies sort_mutate
}
\examples{
library(dplyr)
shapes <- data.frame(length = c(10, 5, 8), width = c(2, 4, 5))
area_shapes <- shapes |> sort_mutate(length*width, desc = TRUE)
area_shapes

# for now it doesn't show the new column, but it arranges according to the calculated area correctly
}
