% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mutate.R
\name{sort_mutate}
\alias{sort_mutate}
\title{Applies sort_mutate}
\usage{
sort_mutate(.data, new_col, expr, desc = FALSE)
}
\arguments{
\item{.data}{dataframe}

\item{new_col}{new column name}

\item{expr}{expression, mathematical formula for your mutate}

\item{desc}{automatically set to FALSE}
}
\value{
a table with mutated new column
}
\description{
Applies sort_mutate
}
\examples{
library(dplyr)
ranked_employees <- employees |>
  sort_mutate(new_col = "earnings", expr = hours_worked * hourly_wage, desc = TRUE)
print(ranked_employees)

}
